+++ projects/core/src/data-providers/sql-database.ts
@@ -22,6 +22,7 @@ import {
   FilterConsumerBridgeToSqlRequest,
   dbNamesOfWithForceSqlExpression,
   isDbReadonly,
+  toDbSql,
 } from '../filter/filter-consumer-bridge-to-sql-request.js'
 import {
   Filter,
@@ -33,6 +34,7 @@ import {
   GroupByOperators,
   type EntityFilter,
   type EntityMetadata,
+  type InsertOrUpdateOptions,
 } from '../remult3/remult3.js'
 import type {
   EntityBase,
@@ -52,6 +54,7 @@ import type {
   MigrationCode,
 } from '../../migrations/migration-types.js'
 import { isOfType } from '../isOfType.js'
+import { originalSqlExpressionKey } from '../filter/fieldDbName.js'
 
 /**
  * A DataProvider for Sql Databases
@@ -384,11 +387,16 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
   async find(options?: EntityDataProviderFindOptions): Promise<any[]> {
     let e = await this.init()
 
-    let { colKeys, select } = this.buildSelect(e, options?.select)
+    let r = this.sql.createCommand()
+    let { colKeys, select } = await this.buildSelect(
+      e,
+      r,
+      options?.select,
+      options?.args,
+    )
     select = 'select ' + select
 
     select += '\n from ' + e.$entityName
-    let r = this.sql.createCommand()
     if (options) {
       if (options.where) {
         let where = new FilterConsumerBridgeToSqlRequest(r, e)
@@ -413,7 +421,9 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
             first = false
           } else select += ', '
 
-          select += e.$dbNameOf(c.field)
+          select += c.field.options.sqlExpression
+            ? c.field.options.key
+            : await e.$dbNameOf(c.field)
           if (c.isDescending) select += ' desc'
           if (this.sql._getSourceSql().orderByNullsFirst) {
             if (c.isDescending) select += ' nulls last'
@@ -457,7 +467,12 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
     return result
   }
 
-  private buildSelect(e: EntityDbNamesBase, selectedFields?: string[]) {
+  private async buildSelect(
+    e: EntityDbNamesBase,
+    r: SqlCommand,
+    selectedFields?: string[],
+    args?: any,
+  ) {
     let select = ''
     let colKeys: FieldMetadata[] = []
     for (const x of this.entity.fields) {
@@ -465,7 +480,15 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
       if (x.isServerExpression) {
       } else {
         if (colKeys.length > 0) select += ', '
-        select += e.$dbNameOf(x)
+        if (typeof x.options.sqlExpression === 'function') {
+          let sql = await (x as any)[originalSqlExpressionKey](
+            this.entity,
+            args,
+            r,
+          )
+          if (sql.includes(' ')) select += '(' + sql + ')'
+          else select += sql
+        } else select += e.$dbNameOf(x)
         if (x.options.sqlExpression) select += ' as ' + x.key
         colKeys.push(x)
       }
@@ -473,7 +496,11 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
     return { colKeys, select }
   }
 
-  async update(id: any, data: any): Promise<any> {
+  async update(
+    id: any,
+    data: any,
+    options?: InsertOrUpdateOptions,
+  ): Promise<any> {
     let e = await this.init()
     let r = this.sql.createCommand()
 
@@ -488,7 +515,7 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
           if (!added) added = true
           else statement += ', '
 
-          statement += e.$dbNameOf(x) + ' = ' + r.param(v)
+          statement += e.$dbNameOf(x) + ' = ' + toDbSql(r, x, v)
         }
       }
     }
@@ -497,10 +524,11 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
     let f = new FilterConsumerBridgeToSqlRequest(r, e)
     Filter.fromEntityFilter(this.entity, idFilter).__applyToConsumer(f)
     statement += await f.resolveWhere()
-    let { colKeys, select } = this.buildSelect(e)
+    let { colKeys, select } = await this.buildSelect(e, r, undefined, undefined)
     let returning = true
     if (this.sql._getSourceSql().doesNotSupportReturningSyntax)
       returning = false
+    if (options?.select === 'none') returning = false
     if (
       returning &&
       this.sql._getSourceSql().doesNotSupportReturningSyntaxOnlyForUpdate
@@ -511,6 +539,7 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
     return r.execute(statement).then((sqlResult) => {
       this.sql._getSourceSql().afterMutation?.()
       if (!returning) {
+        if (options?.select === 'none') return undefined!
         return getRowAfterUpdate(this.entity, this, data, id, 'update')
       }
       if (sqlResult.rows.length != 1)
@@ -538,7 +567,7 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
       this.sql._getSourceSql().afterMutation?.()
     })
   }
-  async insert(data: any): Promise<any> {
+  async insert(data: any, options?: InsertOrUpdateOptions): Promise<any> {
     let e = await this.init()
 
     let r = this.sql.createCommand()
@@ -558,15 +587,18 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
           }
 
           cols += e.$dbNameOf(x)
-          vals += r.param(v)
+          vals += toDbSql(r, x, v)
         }
       }
     }
 
     let statement = `insert into ${e.$entityName} (${cols}) values (${vals})`
 
-    let { colKeys, select } = this.buildSelect(e)
-    if (!this.sql._getSourceSql().doesNotSupportReturningSyntax)
+    let { colKeys, select } = await this.buildSelect(e, r, undefined, undefined)
+    if (
+      !this.sql._getSourceSql().doesNotSupportReturningSyntax &&
+      !(options?.select === 'none')
+    )
       statement += ' returning ' + select
     return await r.execute(statement).then((sql) => {
       this.sql._getSourceSql().afterMutation?.()
@@ -578,15 +610,18 @@ class ActualSQLEntityDataProvider implements EntityDataProvider {
               'Auto increment, for a database that is does not support returning syntax, should return an array with the single last added id. Instead it returned: ' +
                 JSON.stringify(id),
             )
+          if (options?.select === 'none') return undefined!
           return this.find({
             where: new Filter((x) =>
               x.isEqualTo(this.entity.idMetadata.field, id),
             ),
           }).then((r) => r[0])
         } else {
+          if (options?.select === 'none') return undefined!
           return getRowAfterUpdate(this.entity, this, data, undefined, 'insert')
         }
       }
+      if (options?.select === 'none') return undefined!
       return this.buildResultRow(colKeys, sql.rows[0], sql)
     })
   }