+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 All notable changes to this project will be documented in this file.
 
-## [3.2.3]
+## [3.3.0] - 2025-11-09
+
+- [#820](https://github.com/remult/remult/issues/820) - option to eliminate update/insert result to improve performance.
+  ```ts
+  repo(Task).Insert({ id: 1, name: 'test' }, { select: 'none' })
+  ```
+- Added option for `deleteMany({where:'all'})` and `updateMany({where:'all',set:{ status:0 }})`
+- Fixed a bunch if issues around `ValueConverter.DateOnlyString`
+
+## [3.2.3] 2025-10-25
 
 - Fix Backend method issue #890